Mappers and url

+++ app/src/main/java/eu/sesma/paraguas/domain/CurrentWeather.kt
@@ -1,10 +1,10 @@
 package eu.sesma.paraguas.domain
 
 data class CurrentWeather(
-        val precip1hrMetric: Float = 0f,
+        val precip1hrMetric: Double = 0.0,
         val iconUrl: String = "",
         val iconName: String = "",
-        val temp: Float = 0f,
-        val feelsLike: Float = 0f,
+        val temp: Double = 0.0,
+        val feelsLike: Double = 0.0,
         val condition: String = ""
 )
